requirements.txt
yfinance>=0.2.33
python-dotenv>=1.0.0
jinja2>=3.1.2
pandas>=2.0.0
